refactor: rename dep_graph to dep_analysis
