+++ .github/workflows/release.yml
@@ -10,6 +10,7 @@ on:
         options:
           - patch
           - minor
+          - major
         required: true
 
 jobs:
