+++ Cargo.toml
@@ -17,7 +17,8 @@ bundler = ["swc_bundler", "swc_ecma_transforms_optimization", "swc_graph_analyze
 cjs = ["utils", "visit"]
 codegen = ["swc_ecma_codegen", "swc_ecma_codegen_macros", "swc_macros_common"]
 compat = ["transforms", "swc_ecma_transforms_compat", "swc_trace_macro", "swc_config", "swc_config_macro"]
-dep_graph = ["visit"]
+dep_analysis = ["visit"]
+dep_graph = ["swc_ecma_dep_graph"]
 module_specifier = ["deno_media_type/module_specifier", "url"]
 proposal = ["transforms", "swc_ecma_transforms_proposal", "swc_ecma_transforms_classes", "swc_ecma_transforms_macros", "swc_macros_common"]
 react = ["transforms", "swc_ecma_transforms_react", "swc_ecma_transforms_macros", "swc_config", "swc_config_macro", "swc_macros_common"]
@@ -69,6 +70,9 @@ swc_macros_common = { version = "=0.3.8", optional = true }
 swc_trace_macro = { version = "=0.1.3", optional = true }
 swc_visit = { version = "=0.5.7", optional = true }
 swc_visit_macros = { version = "=0.5.8", optional = true }
+
+# todo(dsherret): remove this on the next breaking deno_ast change
+swc_ecma_dep_graph = { version = "=0.113.22", optional = true }
 
 [dev-dependencies]
 pretty_assertions = "1.3.0"
